% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescores_app.R
\name{bayescores}
\alias{bayescores}
\title{Bayescores: Full Bayesian Pipeline (Interactive Shiny App)}
\usage{
bayescores()
}
\description{
Launches an interactive Shiny application that supports end-to-end execution
of the Bayescores Full Bayesian Pipeline, including (i) digitalization of
Kaplan–Meier curves from supplementary materials and (ii) streamlined entry
of quality-of-life (QoL) and toxicity data.
}
