% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esmo_mcbs.R
\name{calculate_esmo_mcbs}
\alias{calculate_esmo_mcbs}
\title{Calculate the ESMO-MCBS v2.0 Score (Guided by Objective)}
\usage{
calculate_esmo_mcbs(objective, ...)
}
\arguments{
\item{objective}{A character string specifying the main goal of the clinical
trial. This is a **mandatory** argument.}

\item{...}{Additional arguments required for the chosen objective. See the
documentation and examples for which arguments are needed for each objective.}
}
\value{
A character string with the final ESMO-MCBS score.
}
\description{
Calculates the ESMO-MCBS v2.0 score based on a primary study
objective. The function first requires the user to specify the study's
main goal, and then uses the corresponding arguments to apply the
correct ESMO form logic.
}
