% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testeo.R, R/toxicity.R
\name{calculate_toxicity_analysis}
\alias{calculate_toxicity_analysis}
\title{A test function}
\usage{
calculate_toxicity_analysis(
  trial_data,
  n_simulations,
  unacceptable_rel_increase = 0.5,
  k_uncertainty = 5
)

calculate_toxicity_analysis(
  trial_data,
  n_simulations,
  unacceptable_rel_increase = 0.5,
  k_uncertainty = 5
)
}
\arguments{
\item{trial_data}{A list containing the toxicity data.}

\item{n_simulations}{An integer for the number of simulations.}

\item{unacceptable_rel_increase}{A number for the unacceptable increase limit.}

\item{k_uncertainty}{A number for the uncertainty constant.}
}
\value{
A list.

A list containing the results.
}
\description{
A simple description.

Calculates weighted toxicity scores (WTS).
}
