% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm649}
\alias{cm649}
\title{Data from the CheckMate-649 Trial}
\format{
A data frame with three variables
\describe{
  \item{time}{A numeric vector representing the survival time.}
  \item{event}{A numeric vector indicating the event status (1 = event, 0 = censored).}
  \item{arm}{A factor or character vector for the treatment arm.}
  \item{id}{Patient id}
}
}
\source{
\url{http://link.to.the.publication/glow}
}
\usage{
cm649
}
\description{
A dataset containing digitized survival data from a clinical trial.
}
\keyword{datasets}
