% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity.R
\name{create_amit_plot}
\alias{create_amit_plot}
\title{Create an AMIT plot (Adverse Event Dot Plot) from a clinical trial object.}
\usage{
create_amit_plot(
  trial_object,
  grade_type = "any_grade",
  main_title = "",
  data_element = "datos",
  n_element = "N_pacientes",
  ...
)
}
\arguments{
\item{trial_object}{The list object containing the trial data.}

\item{grade_type}{A string: "any_grade" (G1–4) or "severe_grade" (G3–4).}

\item{main_title}{The main title of the plot.}

\item{data_element}{The name (as a character string) of the element containing the toxicity data frame.}

\item{n_element}{The name (as a character string) of the element containing the vector with patient counts.}

\item{...}{Additional arguments passed to the HH::AEdotplot function.}
}
\description{
This version is flexible and allows specifying the names of the elements
within the trial object.
}
