% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guyot.R
\name{.place_censors}
\alias{.place_censors}
\title{Helper to Place Censored Observations}
\usage{
.place_censors(a, b, ci, eps = 1e-06, mode = c("even", "uniform"))
}
\arguments{
\item{a}{(`numeric`)\cr Start of the interval.}

\item{b}{(`numeric`)\cr End of the interval.}

\item{ci}{(`integer`)\cr Number of censored observations to place.}

\item{eps}{(`numeric`)\cr A small epsilon value for numerical comparisons.}

\item{mode}{(`character`)\cr Method for placing censored observations, either "even"
(evenly spaced) or "uniform" (from a uniform distribution).}
}
\value{
A numeric vector of censored times.
}
\description{
Distributes a given number of censored observations within a time interval `[a, b]`.
}
\keyword{internal}
