% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esopec_toxicity}
\alias{esopec_toxicity}
\title{Toxicity and QoL data from esopec_toxicity}
\format{
A data frame with 8 variables
\describe{
  \item{EventName}{Adverse event name.}
  \item{SystemOrganClass}{System organ class.}
  \item{Incidence_G1_4_Control}{Incidence (\%) any-grade in control.}
  \item{Incidence_G1_4_Experimental}{Incidence (\%) any-grade in experimental.}
  \item{Incidence_G3_4_Control}{Incidence (\%) grade 3–4 in control.}
  \item{Incidence_G3_4_Experimental}{Incidence (\%) grade 3–4 in experimental.}
  \item{qol}{Numeric QoL score.}
  \item{N_patients}{Named numeric vector with n per arm.}
}
}
\source{
\url{<pon-la-correcta>}
}
\usage{
data(esopec_toxicity)
}
\description{
Toxicity and QoL data from esopec_toxicity
}
