% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summaries.R
\name{extract_mcmc_cure_diffs}
\alias{extract_mcmc_cure_diffs}
\title{Extract MCMC Samples for the Difference in Cure Rates}
\usage{
extract_mcmc_cure_diffs(model)
}
\arguments{
\item{model}{An object of class 'bcm_fit' returned by
`fit_bayesian_cure_model`.}
}
\value{
A numeric vector containing the posterior samples of the difference
  in cure rates (treatment - control).
}
\description{
Extracts the full vector of posterior MCMC samples for the
absolute difference in cure rates between the treatment and control arms.
}
