% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summaries.R
\name{extract_mcmc_time_ratios}
\alias{extract_mcmc_time_ratios}
\title{Extract MCMC Samples for the Time Ratio}
\usage{
extract_mcmc_time_ratios(model)
}
\arguments{
\item{model}{An object of class 'bcm_fit' returned by
`fit_bayesian_cure_model`.}
}
\value{
A numeric vector containing the posterior samples of the Time Ratio.
}
\description{
Extracts the full vector of posterior MCMC samples for the Time
Ratio (TR), allowing for custom plotting and analysis.
}
