% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEYNOTE062}
\alias{KEYNOTE062}
\title{Data from the KEYNOTE062 Trial}
\format{
A data frame with three variables
\describe{
  \item{time}{Numeric vector representing the survival time (e.g., in months).}
  \item{event}{Numeric vector indicating event status (1 for event, 0 for censored).}
  \item{id}{Numeric vector to identify subjetc.}
  \item{arm}{Factor or character vector for treatment arm (e.g., 'Control', 'Experimental').}
}
}
\source{
\url{http://link.to.the.publication/KEYNOTE062}
}
\usage{
KEYNOTE062
}
\description{
A dataset containing digitized survival data from a clinical trial.
}
\keyword{datasets}
