% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kn119}
\alias{kn119}
\title{Digitized survival data from KEYNOTE-119}
\format{
A data frame with five variables:
\describe{
  \item{time}{Numeric; survival time (months).}
  \item{event}{Numeric; 1=event, 0=censored.}
  \item{arm}{Factor/character; treatment arm.}
  \item{id}{Factor/character; patient id.}
  \item{study}{Character; study label.}
}
}
\source{
\url{http://link.to.the.publication/NSABPB40_3B}
}
\usage{
data(kn119)
}
\description{
Digitized survival data from KEYNOTE-119
}
