% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mosaic}
\alias{mosaic}
\title{IPD data from the MOSAIC trial}
\format{
A data frame with 4 variables
\describe{
  \item{time}{Numerical time to event or censoring.}
  \item{status}{Event indicator (1 = event, 0 = censored).}
  \item{curve}{Numeric identifier for the survival curve.}
  \item{arm}{Character string indicating the treatment arm.}
}
}
\source{
\url{onconorte modelos cura/mosaic/ipd.xlsx}
}
\usage{
data(mosaic)
}
\description{
IPD data from the MOSAIC trial
}
