% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_shrinkage.R
\name{outcomes_shrinkage}
\alias{outcomes_shrinkage}
\title{Generate a Summary Table with Optional Empirical Shrinkage}
\usage{
outcomes_shrinkage(
  fit,
  digits = 2,
  empirical_shrinkage = TRUE,
  return_draws = FALSE
)

outcomes_shrinkage(
  fit,
  digits = 2,
  empirical_shrinkage = TRUE,
  return_draws = FALSE
)
}
\arguments{
\item{fit}{The fitted model object from `fit_bayesian_cure_model`.}

\item{digits}{The number of decimal places for rounding in the summary table.}

\item{empirical_shrinkage}{Logical. If TRUE, corrected estimates/draws are generated.}

\item{return_draws}{Logical. If TRUE, the function returns a list of MCMC
draws instead of a summary table. Defaults to FALSE.}
}
\value{
A tibble with key model outcomes. If `empirical_shrinkage` is TRUE,
  includes additional rows with corrected estimates.

If `return_draws` is FALSE (default), a tibble with a summary of outcomes.
  If `return_draws` is TRUE, a list containing vectors of MCMC draws for
  various parameters, including corrected versions if requested.
}
\description{
This function calculates key outcomes from a Bayesian model fit and provides
an option to include empirically shrunk estimates for the Time Ratio and
Odds Ratio based on the method by Van Zwet et al. (2021).

This function calculates key outcomes from a Bayesian model fit. It can either
return a summary table or a list of MCMC draws. It also provides an option
to include empirically shrunk estimates based on Van Zwet et al. (2021).
}
