% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescores.R
\name{plot_bayes_dist}
\alias{plot_bayes_dist}
\title{Plot BayeScores Density Distribution}
\usage{
plot_bayes_dist(scores_list)
}
\arguments{
\item{scores_list}{A list object returned by 'calculate_bayescores'.}
}
\value{
A ggplot object representing the density plot.
}
\description{
Creates a smooth, filled density plot of the posterior distribution of the final scores.
}
