% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescores.R
\name{plot_bayes_donut}
\alias{plot_bayes_donut}
\title{Plot BayeScore Donut Visualization}
\usage{
plot_bayes_donut(scores_list, tr_weight = 0.8)
}
\arguments{
\item{scores_list}{A list object returned by `calculate_bayescores`.}

\item{tr_weight}{The weight assigned to the Time Ratio component in the final score. Default is 0.8.}
}
\value{
A ggplot2 object representing the donut plot.
}
\description{
Generates a donut-shaped visualization of the BayeScore.
}
