% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plotting.R
\name{plot_densities}
\alias{plot_densities}
\alias{plot_densities.bcm_fit}
\title{Plot Posterior Densities}
\usage{
plot_densities(x, ...)

\method{plot_densities}{bcm_fit}(x, ...)
}
\arguments{
\item{x}{The fitted model object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A ggplot object containing the density plots.
}
\description{
A generic function to visualize posterior densities of key parameters
from a Bayesian model fit.
}
