% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity.R
\name{plot_toxicity_adjustment}
\alias{plot_toxicity_adjustment}
\title{Plot the Toxicity Adjustment Distribution}
\usage{
plot_toxicity_adjustment(toxicity_results, trial_name = "Trial")
}
\arguments{
\item{toxicity_results}{A list object returned by `calculate_toxicity_adjustment`.}

\item{trial_name}{A character string for the trial name.}
}
\value{
A ggplot2 object.
}
\description{
Creates a density ridge plot for the toxicity adjustment factor.
}
