% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity.R
\name{plot_toxicity_density}
\alias{plot_toxicity_density}
\title{Plot Toxicity Score Density}
\usage{
plot_toxicity_density(analysis_output)
}
\arguments{
\item{analysis_output}{A list produced by `calculate_toxicity_analysis`.}
}
\value{
A ggplot object.
}
\description{
Creates a modern density plot of the toxicity effect score.
}
