% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescores.R
\name{print.bayes_summary}
\alias{print.bayes_summary}
\title{Print method for bayes_summary objects}
\usage{
\method{print}{bayes_summary}(x, ...)
}
\arguments{
\item{x}{An object of class 'bayes_summary'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for bayes_summary objects
}
