% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{run_plot_scenario}
\alias{run_plot_scenario}
\title{Run a Scenario for a Sensitivity Plot}
\usage{
run_plot_scenario(
  x_var_name,
  x_range,
  line_var_name,
  line_levels,
  fixed_params,
  cal_args
)
}
\arguments{
\item{x_var_name}{The parameter to vary on the x-axis.}

\item{x_range}{The range for the x-axis variable.}

\item{line_var_name}{The parameter to vary as different lines.}

\item{line_levels}{The levels for the line variable.}

\item{fixed_params}{A list of fixed parameters.}

\item{cal_args}{A list of calibration arguments.}
}
\value{
A data frame with simulation results.
}
\description{
Generates the data needed for a single sensitivity plot line chart.
}
