% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescores.R
\name{summarize_bayescores}
\alias{summarize_bayescores}
\title{Summarize BayeScores Results}
\usage{
summarize_bayescores(scores_list)
}
\arguments{
\item{scores_list}{A list object returned by 'calculate_bayescores'.}
}
\value{
An object of class 'bayes_summary' containing the median, 95% CrI,
  and the rescaled score.
}
\description{
Calculates summary statistics for final BayeScores. This function returns an
object of class 'bayes_summary'. The summary is printed to the console
when the object is called.
}
