% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summaries.R
\name{summarize_cure_rates}
\alias{summarize_cure_rates}
\title{Summarize Cure Rate Results}
\usage{
summarize_cure_rates(model)
}
\arguments{
\item{model}{An object of class 'bcm_fit' returned by
`fit_bayesian_cure_model`.}
}
\value{
A list containing the estimated cure rates for each arm (named
  dynamically using the factor levels), their absolute difference, and an
  `interpretation` string.
}
\description{
Calculates the posterior estimates of the cure rates for each
treatment arm and the absolute difference between them.
}
