% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_shrinkage.R
\name{get_bayescores_draws}
\alias{get_bayescores_draws}
\title{Get Original or Shrunk MCMC Draws for Subsequent Analyses}
\usage{
get_bayescores_draws(fit, shrinkage_method = "none")
}
\arguments{
\item{fit}{The fitted model object from `fit_bayesian_cure_model`.}

\item{shrinkage_method}{A character string specifying the shrinkage method.
Options are "zwet", "sherry", or "none" (default).}
}
\value{
A list containing two named vectors of posterior samples:
  \itemize{
    \item{\code{tr_posterior_samples}: Draws for the Time Ratio (already exponentiated).}
    \item{\code{cure_posterior_samples}: Draws for the absolute difference
      in cure rates.}
  }
}
\description{
Extracts or generates posterior draws for efficacy inputs, which can be used
in other functions like `get_bayescores`. It can return the original MCMC draws
or generate new draws corrected with an empirical shrinkage method.
}
