% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Plot Model Diagnostics}
\usage{
model_diagnostics(x, ...)
}
\arguments{
\item{x}{An object, typically a model fit.}

\item{...}{Other arguments.
• @export}
}
\description{
A generic function to plot model diagnostics.
}
