% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summaries.R
\name{outcomes}
\alias{outcomes}
\title{Generates a summary table of outcomes from a Bayesian model}
\usage{
outcomes(fit, digits = 2)
}
\arguments{
\item{fit}{The fitted model object, containing 'stan_fit'.}

\item{digits}{The number of decimal places to round the results.}
}
\value{
A tibble with key outcomes: Time Ratio, Odds Ratio,
        and long-term survival rates for each arm.
}
\description{
Generates a summary table of outcomes from a Bayesian model
}
