% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_shrinkage.R
\name{outcomes}
\alias{outcomes}
\title{Generate a Summary Table of Model Outcomes}
\usage{
outcomes(fit, digits = 2, shrinkage_method = "none")
}
\arguments{
\item{fit}{The fitted model object from `fit_bayesian_cure_model`.}

\item{digits}{The number of decimal places for rounding.}

\item{shrinkage_method}{A character string specifying the shrinkage method for
the corrected estimates. Options are "zwet", "sherry", or "none" (default).}
}
\value{
A tibble with a summary of key model outcomes.
}
\description{
Calculates key outcomes from a Bayesian model fit and can display
side-by-side comparisons with empirically shrunk estimates.
}
